+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - execute list repos, list organizations and create ticket always in kidney workspace
 - Error if ticket already exists
 - Successfully open VSCode on Windows
+- Rename GitCloner to GitHandler
 
 ### Removed